Move the polybasis to cons_law
The grid should not depend at all on the polynomial basis
selected. I moved to the conservation law object definition.
However, each one of the cells of the grid must have the same
definition of coefficients for the polynomial basis.

+++ utilities.h
@@ -91,6 +91,6 @@ std::vector<std::vector<Number>> transpose(std::vector<std::vector<Number>>& );
 //std::vector<std::vector<double> > operator|= (std::vector<std::vector<double> > & );
 
 //template<typename Number>
-std::vector<std::vector<long double> > LegendreCoefficients(int);
+std::vector<std::vector<double> > LegendreCoefficients(int);
